+++ src/resources/views/images.blade.php
@@ -14,26 +14,38 @@
     </form>
 </div>
 
-<div class="container image-container">
+<div class="container">
     <hr />
     <h2>Image Library</h2>
     <p>Copy the image filename below and paste into the image field when creating/editing a post.</p>
 
-    <div class="row">
-        <?php $images_path = array_diff(File::allFiles(public_path("storage/images/posts"), array('.', '..'))); ?>
-        @foreach($images_path as $image)
+    @php $images = File::allFiles(public_path("storage/images/posts")); @endphp
+    @foreach(array_values($images) as $index => $image)
         @php $image_name = basename($image); @endphp
-        <div class="col">
-            <img src='{{ asset("storage/images/posts/$image_name") }}' class="img-thumbnail">
+        @if ($index == 0)
+            <div class="row image-row-container">
+        @endif
+
+        <div class="col text-center">
+            <div class="flex-center-container">
+                <img src='{{ asset("storage/images/posts/$image_name") }}' class="img-thumbnail image-preview">
+            </div>
             <p>{{ $image_name }}</p>
             <form action="{{ route('delete-image') }}" method="post">
                 @csrf
                 <input type="hidden" name="id" value="{{ $image_name }}">
-                <input type="submit" value="Delete Image" class="btn btn-sm btn-danger inline-block">
+                <a class="btn btn-primary btn-sm" href="{{ asset("storage/images/posts/$image_name") }}" download>Download</a>
+                <input type="submit" value="Delete" class="btn btn-sm btn-danger">
             </form>
         </div>
-        @endforeach
-    </div>
+
+        @if (($index + 1) % 3 == 0)
+            </div>
+            <div class="row image-row-container">
+        @elseif ($index == count($images) - 1)
+            </div>
+        @endif
+    @endforeach
 </div>
 
 @endsection